physical-ai-robotics-textbook/translation/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
markdown==3.5.1
PyYAML==6.0.1
redis==5.0.1  # For caching
python-jose[cryptography]==3.3.0  # For JWT handling
passlib[bcrypt]==1.7.4  # For password hashing
python-dotenv==1.0.0  # For environment variables
requests==2.31.0
asyncio==3.4.3